enemy update finished
