+++ Game/enemy/property.c
@@ -10,6 +10,9 @@ void Enemy_Init(struct Enemy *enemy,struct Object* body,struct Object* face, str
 	enemy->Critical_Rate = 50;//critical damage to player; should not be higher than 50
 
 	Transform_Init(&enemy->transform, NULL);
+    Transform_Init(&enemy->face.transform, &enemy->transform);
+    Transform_Init(&enemy->body.transform, &enemy->transform);
+    Transform_Init(&enemy->leg.transform, &enemy->transform);
 
 	struct Vector3 v1, v2;
 	Vector3_Set(&v1, 1, -1, 1);
@@ -27,17 +30,3 @@ void Enemy_Start(struct Enemy *enemy)
 	//TODO enemy start
 }
 
-void Enemy_Update(struct Enemy* enemy, bool do_find_way, struct Scene* scene, struct link* current)
-{
-	if (do_find_way){
-		struct link* beginning = Find_Way(&enemy->body,scene);//beginning actually refers to current position of bot.
-		current = beginning;
-	}
-	struct Transform* target = current->next->current;
-	struct Vector3* move = (struct Vector3*)malloc(sizeof(struct Vector3));
-	Vector3_Set(move,
-				target->globalPosition.x - enemy->transform.globalPosition.x,
-				target->globalPosition.y - enemy->transform.globalPosition.y,
-				target->globalPosition.z - enemy->transform.globalPosition.z);
-	Enemy_Move(enemy, move);
-}
