+++ Game/enemy/property.h
@@ -3,10 +3,7 @@
 
 #include "../../op_engine/op_engine.h"
 #include "../player.h"
-#include "../scene.h"
-#include "behaviour.h"
-#include "damage.h"
-#include "findway.h"
+
 
 struct Enemy{
     struct Object face,leg,body;
@@ -22,6 +19,6 @@ void Del_Enemy(struct Enemy *enemy);
 
 void Enemy_Start(struct Enemy *enemy);
 
-void Enemy_Update(struct Enemy *enemy, bool do_find_way, struct Scene* scene, struct link* current);
+
 
 #endif //FPS_SIMULATOR_PROPERTY_H
